fix: изменить аннотации для entity классов; refactor: добавить пакеты для всех классов и интерфесов; style: убрать лишние строчки и комментарии
