+++ ewm-service/src/main/java/ru/practicum/explorewm/category/mapper/CategoryMapper.java
@@ -1,4 +1,4 @@
-package ru.practicum.explorewm.category;
+package ru.practicum.explorewm.category.mapper;
 
 import ru.practicum.explorewm.category.dto.CategoryDto;
 import ru.practicum.explorewm.category.dto.CreateCategory;
@@ -9,14 +9,12 @@ import java.util.Collection;
 
 public class CategoryMapper {
 
-    // из create в entity
     public static Category mapToCategory(CreateCategory request) {
         Category category = new Category();
         category.setName(request.getName());
         return category;
     }
 
-    // из dto в entity
     public static Category mapToCategory(CategoryDto request) {
         Category category = new Category();
         category.setId(request.getId());
@@ -24,7 +22,6 @@ public class CategoryMapper {
         return category;
     }
 
-    // из entity в dto
     public static CategoryDto mapToCategoryDto(Category category) {
         CategoryDto response = new CategoryDto();
         response.setId(category.getId());
@@ -32,7 +29,6 @@ public class CategoryMapper {
         return response;
     }
 
-    // из iterable в dto
     public static Collection<CategoryDto> mapToLisCategoryDto(Iterable<Category> categories) {
         Collection<CategoryDto> response = new ArrayList<>();
         for (Category category : categories) {
