+++ ewm-service/src/main/java/ru/practicum/explorewm/category/model/Category.java
@@ -1,16 +1,15 @@
 package ru.practicum.explorewm.category.model;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import javax.persistence.*;
 
 @Entity
-@Data
+@Getter
+@Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@Table(name = "categories", schema = "public",  uniqueConstraints = @UniqueConstraint(columnNames = "name"))
+@Table(name = "categories", uniqueConstraints = @UniqueConstraint(columnNames = "name"))
 public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
