+++ ewm-service/src/main/java/ru/practicum/explorewm/compilation/service/CompilationServiceImpl.java
@@ -5,8 +5,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.explorewm.compilation.CompilationMapper;
-import ru.practicum.explorewm.compilation.CompilationRepository;
+import ru.practicum.explorewm.compilation.mapper.CompilationMapper;
+import ru.practicum.explorewm.compilation.repository.CompilationRepository;
 import ru.practicum.explorewm.compilation.dto.CompilationDto;
 import ru.practicum.explorewm.compilation.dto.CompilationCreate;
 import ru.practicum.explorewm.compilation.model.Compilation;
