+++ ewm-service/src/main/java/ru/practicum/explorewm/event/controller/PrivateEventController.java
@@ -3,7 +3,7 @@ package ru.practicum.explorewm.event.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.explorewm.event.EventMapper;
+import ru.practicum.explorewm.event.mapper.EventMapper;
 import ru.practicum.explorewm.event.dto.*;
 import ru.practicum.explorewm.event.service.EventService;
 import ru.practicum.explorewm.requests.dto.ParticipationRequestDto;
