+++ ewm-service/src/main/java/ru/practicum/explorewm/event/dto/EventCreate.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.explorewm.CustomDateTimeDeserializer;
-import ru.practicum.explorewm.event.InTwoHours;
+import ru.practicum.explorewm.dataio.CustomDateTimeDeserializer;
+import ru.practicum.explorewm.event.validator.InTwoHours;
 import ru.practicum.explorewm.event.location.model.Location;
 
 import javax.validation.constraints.Future;
@@ -17,35 +17,24 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class EventCreate {
-
     @Size(min = 20, max = 2000)
     private String annotation;
-
     @NotNull
     private Long category;
-
     @Size(min = 20, max = 7000)
     private String description;
-
     @Future
     @InTwoHours
     @JsonDeserialize(using = CustomDateTimeDeserializer.class)
     private LocalDateTime eventDate;
-
     @NotNull
     private Location location;
-
     @NotNull
     private Boolean paid;
-
     @NotNull
     private Integer participantLimit;
-
     @NotNull
     private Boolean requestModeration;
-
     @Size(min = 3, max = 120)
     private String title;
-
-
 }
