+++ ewm-service/src/main/java/ru/practicum/explorewm/event/dto/EventFull.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.explorewm.CustomDateTimeSerializer;
+import ru.practicum.explorewm.dataio.CustomDateTimeSerializer;
 import ru.practicum.explorewm.category.model.Category;
-import ru.practicum.explorewm.event.State;
+import ru.practicum.explorewm.event.state.State;
 import ru.practicum.explorewm.event.location.dto.LocationDto;
 import ru.practicum.explorewm.user.dto.UserShort;
