+++ ewm-service/src/main/java/ru/practicum/explorewm/event/dto/EventPublished.java
@@ -3,9 +3,9 @@ package ru.practicum.explorewm.event.dto;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import ru.practicum.explorewm.CustomDateTimeSerializer;
+import ru.practicum.explorewm.dataio.CustomDateTimeSerializer;
 import ru.practicum.explorewm.category.model.Category;
-import ru.practicum.explorewm.event.State;
+import ru.practicum.explorewm.event.state.State;
 import ru.practicum.explorewm.event.location.dto.LocationDto;
 import ru.practicum.explorewm.user.dto.UserShort;
 
@@ -31,6 +31,4 @@ public class EventPublished {
     private Boolean requestModeration;
     private State state;
     private String title;
-
-
 }
