+++ ewm-service/src/main/java/ru/practicum/explorewm/event/dto/SortRequest.java
@@ -1,4 +1,4 @@
-package ru.practicum.explorewm.event;
+package ru.practicum.explorewm.event.dto;
 
 public enum SortRequest {
     EVENT_DATE, VIEWS
