+++ ewm-service/src/main/java/ru/practicum/explorewm/event/model/Event.java
@@ -1,12 +1,10 @@
 package ru.practicum.explorewm.event.model;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import ru.practicum.explorewm.category.model.Category;
 import ru.practicum.explorewm.compilation.model.Compilation;
-import ru.practicum.explorewm.event.State;
 import ru.practicum.explorewm.event.location.model.Location;
+import ru.practicum.explorewm.event.state.State;
 import ru.practicum.explorewm.user.model.User;
 
 import javax.persistence.*;
@@ -14,65 +12,48 @@ import java.time.LocalDateTime;
 import java.util.Collection;
 
 @Entity
-@Data
+@Getter
+@Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@Table(name = "events", schema = "public")
+@Table(name = "events")
 public class Event {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @Column
+    @Column(length = 2000)
     private String annotation;
-
     @ManyToOne
     @JoinColumn(name = "category_id", nullable = false)
     private Category category;
-
     @Column(name = "confirmed_requests")
     private int confirmedRequests;
-
     @Column(name = "created_on", nullable = false)
     private LocalDateTime createdOn;
-
-    @Column
+    @Column(length = 7000)
     private String description;
-
     @Column(name = "event_date", nullable = false)
     private LocalDateTime eventDate;
-
     @ManyToOne
     @JoinColumn(name = "initiator_id", nullable = false)
     private User initiator;
-
     @ManyToOne
     @JoinColumn(name = "location_id ", nullable = false)
     private Location location;
-
     @Column
     private Boolean paid;
-
     @Column(name = "participant_limit", nullable = false)
     private int participantLimit;
-
     @Column(name = "published_on")
     private LocalDateTime publishedOn;
-
     @Column(name = "request_moderation", nullable = false)
     private Boolean requestModeration;
-
-    @Column
     @Enumerated(EnumType.STRING)
     private State state;
-
-    @Column
+    @Column(length = 120)
     private String title;
-
     @Column
     private int views;
-
-   @ManyToMany(mappedBy = "events")
+    @ManyToMany(mappedBy = "events")
     private Collection<Compilation> compilations;
-
 }
