+++ ewm-service/src/main/java/ru/practicum/explorewm/event/service/EventServiceImpl.java
@@ -6,11 +6,11 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.explorewm.category.CategoryRepository;
+import ru.practicum.explorewm.category.repository.CategoryRepository;
 import ru.practicum.explorewm.category.model.Category;
-import ru.practicum.explorewm.event.EventMapper;
-import ru.practicum.explorewm.event.SortRequest;
-import ru.practicum.explorewm.event.State;
+import ru.practicum.explorewm.event.mapper.EventMapper;
+import ru.practicum.explorewm.event.dto.SortRequest;
+import ru.practicum.explorewm.event.state.State;
 import ru.practicum.explorewm.event.dto.*;
 import ru.practicum.explorewm.event.location.model.Location;
 import ru.practicum.explorewm.event.model.Event;
@@ -19,9 +19,9 @@ import ru.practicum.explorewm.event.repository.EventRepository;
 import ru.practicum.explorewm.event.repository.LocationRepository;
 import ru.practicum.explorewm.exception.ConditionsNotMetException;
 import ru.practicum.explorewm.exception.ObjectNotFoundException;
-import ru.practicum.explorewm.statistics.SendingStatisticsService;
+import ru.practicum.explorewm.statistics.service.SendingStatisticsService;
 import ru.practicum.explorewm.statistics.dto.EndpointHit;
-import ru.practicum.explorewm.user.UserRepository;
+import ru.practicum.explorewm.user.repository.UserRepository;
 import ru.practicum.explorewm.user.model.User;
 
 import javax.servlet.http.HttpServletRequest;
