+++ ewm-service/src/main/java/ru/practicum/explorewm/requests/controller/PrivateParticipationRequestController.java
@@ -36,5 +36,4 @@ public class PrivateParticipationRequestController {
 
         return service.cancel(userId, requestId);
     }
-
 }
