+++ ewm-service/src/main/java/ru/practicum/explorewm/requests/status/Status.java
@@ -1,6 +1,5 @@
-package ru.practicum.explorewm.requests;
+package ru.practicum.explorewm.requests.status;
 
 public enum Status {
-
     PENDING, CONFIRMED, CANCELED, REJECTED
 }
