+++ ewm-service/src/main/java/ru/practicum/explorewm/user/mapper/UserMapper.java
@@ -1,4 +1,4 @@
-package ru.practicum.explorewm.user;
+package ru.practicum.explorewm.user.mapper;
 
 import ru.practicum.explorewm.user.dto.UserCreate;
 import ru.practicum.explorewm.user.dto.UserShort;
@@ -7,7 +7,6 @@ import ru.practicum.explorewm.user.model.User;
 
 public class UserMapper {
 
-    // из create в entity
     public static User mapToUser(UserCreate request) {
         User user = new User();
         user.setName(request.getName());
@@ -15,12 +14,10 @@ public class UserMapper {
         return user;
     }
 
-    // из entity в short
     public static UserShort mapToUserShort(User user) {
         UserShort response = new UserShort();
         response.setId(user.getId());
         response.setName(user.getName());
         return response;
     }
-
 }
