+++ ewm-service/src/main/java/ru/practicum/explorewm/user/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package ru.practicum.explorewm.user;
+package ru.practicum.explorewm.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.explorewm.user.model.User;
