+++ stats-server/src/main/java/ru/practicum/explorewithme/repository/StatisticsRepository.java
@@ -1,4 +1,4 @@
-package ru.practicum.explorewithme;
+package ru.practicum.explorewithme.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -21,16 +21,15 @@ public interface StatisticsRepository extends JpaRepository<EndpointHit, Long> {
             "count(distinct eh.ip)) " +
             "from EndpointHit as eh " +
             "where eh.uri = ?1 " +
-            "and eh.timestamp < ?2 " +
-            "and eh.timestamp > ?3 group by eh.app")
+            "and eh.timestamp between ?2 and ?3 " +
+            "group by eh.app")
     Collection<StatsCount> getUniqueViews(String uris, LocalDateTime start, LocalDateTime end);
 
     @Query(value = "select new ru.practicum.explorewithme.dto.StatsCount(eh.app, " +
             "count(eh.id)) " +
             "from EndpointHit as eh " +
             "where eh.uri = ?1 " +
-            "and eh.timestamp > ?2 " +
-            "and eh.timestamp < ?3 " +
+            "and eh.timestamp between ?2 and ?3 " +
             "group by eh.app")
     Collection<StatsCount> getViews(String uris, LocalDateTime start, LocalDateTime end);
 
@@ -40,5 +39,4 @@ public interface StatisticsRepository extends JpaRepository<EndpointHit, Long> {
             "where eh.uri = ?1 " +
             "group by eh.app")
     Collection<StatsCount> getViews(String uris);
-
 }
